requirements.txt
mediapipe>=0.10.21
opencv-python>=4.8.0
mido>=1.2.10
python-rtmidi>=1.4.9
numpy>=1.24.0,<2.0.0
Pillow>=9.0.0